Cleaner interfaces between processor and kolme (closes #20)
The main concept I've introduced here is the .read() and .write() methods, which give back a KolmeRead and KolmeWrite, plus the KolmeWriteDb database transaction concept. I don't like the coding style around the database stuff at all, it's abusing some Mutex stuff to work around lifetime issues, but this is good enough for now.

With these changes in place, we can begin to have a clean interface that all components can use for talking to Kolme. Instead of building out the rest of the interface now, I'll add in necessary pieces with upcoming PRs.