Remove transactions from mempool when picked up by processor
This was an oversight in the original code, which would result in failing transactions blocking up the mempool indefinitely. Discovered by @anakinzhed while working on #190. This is a general improvement that also happens to fix that case.

Suspicion of why we had stalled tests in #190: the processor ended up busy-looping hard enough that it starved other tasks and prevented them from continuing with their clean-up activities.