Gossip: make network discovery more reliable.
I experienced some stalled test runs locally because gossip never discovered that it was connected to peers correctly. I don't quite know how that happened with the old code, but the new version is more in line with what we wanted anyway: it actually sends a message to see if the network is correctly connected.