Merkle serialization includes versions #296
This introduces a split between MerkleSerialize and MerkleSerializeRaw,
and a similar split on the deserialize side. MerkleSerialize includes
version information, and a blanket impl for MerkleSerializeRaw automatically
inserts that version information into the serialized payload.

Types that may change their format in the future should implement MerkleSerialize.
Types that have a fixed format, like primitives, can use MerkleSerializeRaw
to bypass the version storage overhead.

Changes here were made to minimize the impact on downstream libraries and help
them "do the right thing" out of the box, by defaulting to using versioning
and automatically setting the version to 0 if unspecified.