Track subhashes for merkle stores (fixes #307)
Previously, we discovered subhashes by iteratively trying to parse payloads. This, however, was O(n^2) in complexity of the size of the payload. Instead, we now store both the payload and the hashes of all the direct children. This allows the merkle-map crate to directly load up all necessary paylods immediately.

The code is a bit of a mess, the goal was to get something out the door ASAP for testing and to confirm that it resolves the current issues faced during an app deployment.