Kademlia tests use in-memory store
Currently, we're using a static directory, which (1) pollutes the working tree and (2) leads to conflicts when changing serialization format. We could use a temp dir instead, but the reality is that we don't need to test Fjall as part of the Kademlia tests, so moving to in-memory makes more sense.