Version upgrade logic (fixes #43)
This addresses the majority of the version upgrade logic, and introduces sufficient data structure additions that we can further refine the upgrade process without needing further data migrations in the future. This does not yet include an update to the docs on version upgrades, that will follow in a separate commit to allow for immediate code review.