Not going to bother checking the hashes (closes #116)
Motivation: we're already implicitly testing all aspects of this already. Adding in this kind of testing would require significant modifications to either the bounds in the public API, or a lot of code duplication which is more likely to introduce issues in the future. Closing the issue as wontfix and removing the FIXMEs.