feat: Incremental state sync
Fixes #237. Fixes #347.

Instead of transferring the entirety of the state contents in one payload, this switches the state sync mechanism in Gossip to transfer individual Merkle layers, avoiding a buffer size limitation and removing redundant data transfers, making state sync more scalable and efficient.

This ended up requiring a far more complex state machine than I initially anticipated. I've included a semi-detailed design doc for this for the docs site, which is a good starting point for understanding this code.