feat: Implement a Trigger mechanism.
A common pattern in the codebase is setting up asynchronous triggers and listeners. Previously, this was all ad-hoc, with confusing code sprinkled everywhere. Instead, we now have a much more explicit Trigger data type and well-named functions. Additionally, by labeling the triggers, we can get useful error messages if the invariants are violated (namely: a Trigger is dropped fully while a TriggerListener is still active).

BREAKING CHANGE: This slightly modifies the public API of Kolme, exposing TriggerListener values instead of tokio::sync::watcher::Receiver for subscriptions. The primary change to implement is to use the .listen() method instead of .changed().