build: Use pre-built localosmosis image for tests
This change simplifies the setup for integration tests by switching the
`localosmosis` service from a local Dockerfile build to using a pre-
built image. This eliminates the need for local build context,
volumes, and custom setup scripts, ensuring a more consistent and
faster environment for running tests.

Additionally, both `localosmosis` and `postgres` services are now
explicitly connected to a new, named `app_net` bridge network. The
reason for this is because without this it didn't work in my
machine (could be a NixOS thing).