Remove buggy/unneeded reverse block map
The idea had been to handle a bunch of pending blocks. However, we're squarely sticking to one block at a time. We may end up with some wasted network traffic in corner cases (like we began syncing one block and then switched to an earlier one), but that's worth the savings in complexity.