ci: Enforce warnings and cache keys
The CI workflow is updated to enforce stricter compilation for Rust
projects. By adding the `RUSTFLAGS: "-D warnings"`, all Rust warnings
will now be treated as errors. The primary motivation is to reduce the
CI time: We pass this option during the clippy step which causes more
work.

*   Add `RUSTFLAGS: "-D warnings"` to treat all Rust warnings as errors.
*   Update `rust-cache` prefix keys for `kolme` and `solana`
workspaces so that fresh cache is generate. I believe the current
cache size is bigger than it's supposed to be because of my various
experiments over the last week.