fix: merkle-map resilience to LRU evictions
This is a hacky fix to the underlying problem. The problem is that, previously, we were using the MerkleManager cache not just as an optimization, but as a bridge between async and sync data loading. This means that the data deserialization mechanism requires that the cache contains all layers. With the addition of using LRU with evictions, this requirements is no longer guaranteed to hold.

This is an initial fix to get the immediate bug resolved. However, the code is cumbersome at this point. In particular, using a MerkleManager no longer seems like a good API design, and instead each Merkle store could optionally include its own LRU-powered in-memory cache just for optimizations.