feat: StorableBlock is lean
StorableBlock initially contained framework state, app state, and logs directly in its serialized payload. We fixed this with version 1 of the data storage, pushing those values into their own layers for better data management in Gossip.

This change goes a step further: StorableBlock explicitly does _not_ contain the data itself. Instead, it only contains the block metadata itself. It's now the responsibility of other parts of the code to load and save the block state values.

The reason for this is performance: often we need only a subset of the data for a block, and needing to load and store the entirety of the block state for each operation was prohibitively expensive.