feat: Implement basic healthz API endpoint
This commit introduces a new `/healthz` endpoint to the Kolme API
server. This simple endpoint provides a basic liveness check,
returning a static "Healthy!" response. It is intended to be used by
load balancers to verify the application's responsiveness.

Having a more sophisticated health check pattern (based on Kolme's is
the next thing), but this will unblock me currently.