perf: add an explicit contents cache
The current cache in merkle-map caches both the contents data and the full value. Due to how it's implemented, it can only be triggered from within Lockable. This reduces the reusability of the cache significantly.

This commit adds a new contents-only caching layer. The goal is to make it unnecessary to have such redundant caching occurring within kolme-store. A follow-up commit will attempt removing that to ensure that everything continues to behave as expected.

One thing to note: this change means that only values _still being used elsewhere_ remain in the cache. By contrast, with the LruCache within kolme-store, values will remain cached until evicted. Due to standard use cases within Kolme, this is probably the right decision. However, we may elect in the future to add additional lru-based caching within contents_cache.