RFC: Clarify versioning policy
This PR is more asking about feedback before I continue my
implementation in https://github.com/fpco/kolme/pull/471

The scheme documented here is used for our client and I have found it
helpful in practice.

Note that the Tooling Compatiblity doesn't exist yet and it's purely
hypothetical. I'm planning to implement it if we agree that it's a
good idea.